requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
astor==0.8.1
    # via flake8-qgis
attrs==22.2.0
    # via
    #   flake8-annotations
    #   flake8-bugbear
    #   pytest
black==23.3.0
    # via -r requirements-dev.in
build==0.10.0
    # via pip-tools
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via
    #   black
    #   pip-tools
coverage[toml]==7.2.3
    # via pytest-cov
debugpy==1.6.7
    # via -r requirements-dev.in
distlib==0.3.6
    # via virtualenv
exceptiongroup==1.1.1
    # via pytest
filelock==3.11.0
    # via virtualenv
flake8==6.0.0
    # via
    #   -r requirements-dev.in
    #   flake8-annotations
    #   flake8-bugbear
    #   flake8-qgis
    #   pep8-naming
flake8-annotations==3.0.0
    # via -r requirements-dev.in
flake8-bugbear==23.3.23
    # via -r requirements-dev.in
flake8-qgis==1.0.0
    # via -r requirements-dev.in
identify==2.5.22
    # via pre-commit
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements-dev.in
mccabe==0.7.0
    # via flake8
mypy==1.2.0
    # via -r requirements-dev.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.7.0
    # via pre-commit
packaging==23.0
    # via
    #   black
    #   build
    #   pytest
pathspec==0.11.1
    # via black
pep8-naming==0.13.3
    # via -r requirements-dev.in
pip-tools==6.13.0
    # via -r requirements-dev.in
platformdirs==3.2.0
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==3.2.2
    # via -r requirements-dev.in
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pyproject-hooks==1.0.0
    # via build
pyqt5-stubs==5.15.6.0
    # via -r requirements-dev.in
pytest==7.2.2
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-qgis
pytest-cov==4.0.0
    # via -r requirements-dev.in
pytest-qgis==1.3.2
    # via -r requirements-dev.in
pyyaml==6.0
    # via pre-commit
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   mypy
    #   pytest
typing-extensions==4.5.0
    # via mypy
virtualenv==20.21.0
    # via pre-commit
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
